split tests into a package, resize/resampling tests into their own file
