+++ pymaging/tests/test_basic.py
@@ -59,27 +59,27 @@ class BasicTests(PymagingBaseTestCase):
             [Green, Blue, Red],
             [Blue, Red, Green],
         ])
-        
+
     def test_open_invalid_image(self):
         self.assertRaises(FormatNotSupported, Image.open, StringIO(''))
 
     def test_crop(self):
         img = self._get_fake_image()
         img.crop(1, 1, 1, 1)
-        
+
     def test_flip_left_right(self):
         img = self._get_fake_image()
         img.flip_left_right()
-        
+
     def test_flip_top_bottom(self):
         img = self._get_fake_image()
         img.flip_top_bottom()
-    
+
     def test_get_pixel(self):
         img = self._get_fake_image()
         color = img.get_color(0, 0)
         self.assertEqual(color, Red)
-        
+
     def test_set_pixel(self):
         img = image_factory([
             [Black, Black],
@@ -90,32 +90,18 @@ class BasicTests(PymagingBaseTestCase):
             [White, Black],
             [Black, Black],
         ])
-        
+
     def test_color_mix_with(self):
         base = Red
         color = Lime.get_for_brightness(0.5)
         result = base.cover_with(color)
         self.assertEqual(result, Color(128, 127, 0, 255))
-        
+
     def test_color_mix_with_fastpath(self):
         base = Red
         color = Lime
         result = base.cover_with(color)
         self.assertEqual(result, Lime)
-
-
-class ResizeCropTests(PymagingBaseTestCase):
-    def test_resize_nearest_resampling(self):
-        img = image_factory([
-            [Red, Green, Blue],
-            [Green, Blue, Red],
-            [Blue, Red, Green],
-        ])
-        img = img.resize(2, 2)
-        self.assertImage(img, [
-            [Red, Blue],
-            [Blue, Green],
-        ])
 
 
 class DrawTests(PymagingBaseTestCase):
@@ -130,14 +116,14 @@ class DrawTests(PymagingBaseTestCase):
             [White, Black],
             [Black, Black],
         ])
-        
+
     def test_alpha_mixing(self):
         img = image_factory([[Red]])
         semi_transparent_green = Lime.get_for_brightness(0.5)
         img.draw(Pixel(0, 0), semi_transparent_green)
         result = img.get_color(0, 0)
         self.assertEqual(result, Color(128, 127, 0, 255))
-        
+
     def test_draw_line_topleft_bottomright(self):
         img = image_factory([
             [Black, Black, Black, Black, Black],
@@ -155,7 +141,7 @@ class DrawTests(PymagingBaseTestCase):
             [Black, Black, Black, White, Black],
             [Black, Black, Black, Black, White],
         ])
-        
+
     def test_draw_line_bottomright_topleft(self):
         img = image_factory([
             [Black, Black, Black, Black, Black],
@@ -173,7 +159,7 @@ class DrawTests(PymagingBaseTestCase):
             [Black, Black, Black, White, Black],
             [Black, Black, Black, Black, White],
         ])
-        
+
     def test_draw_line_bottomleft_topright(self):
         img = image_factory([
             [Black, Black, Black, Black, Black],
@@ -191,7 +177,7 @@ class DrawTests(PymagingBaseTestCase):
             [Black, White, Black, Black, Black],
             [White, Black, Black, Black, Black],
         ])
-        
+
     def test_draw_line_topright_bottomleft(self):
         img = image_factory([
             [Black, Black, Black, Black, Black],
@@ -209,7 +195,7 @@ class DrawTests(PymagingBaseTestCase):
             [Black, White, Black, Black, Black],
             [White, Black, Black, Black, Black],
         ])
-        
+
     def test_draw_line_steep(self):
         img = image_factory([
             [Black, Black, Black, Black, Black],
